use usize for address

+++ src/lib.rs
@@ -2,9 +2,9 @@ extern crate conv;
 pub use conv::conversion::*;
 use std::any::Any;
 pub trait Interconnect: InterconnectClone{
-    fn load8(&self, addr:u32)->Option<u8>;
-    fn store8(&mut self, addr:u32,val:u8);
-    fn load16(&self, addr:u32)->Option<u16>{
+    fn load8(&self, addr:usize)->Option<u8>;
+    fn store8(&mut self, addr:usize,val:u8);
+    fn load16(&self, addr:usize)->Option<u16>{
         if let Some(b0) = self.load8(addr+0){
             if let Some(b1) = self.load8(addr+1){
                return Some(u8_to_u16(&[b0,b1])); 
@@ -12,13 +12,13 @@ pub trait Interconnect: InterconnectClone{
         }
         None
     }
-    fn store16(&mut self, addr:u32,val:u16){
+    fn store16(&mut self, addr:usize,val:u16){
         let v = u16_to_u8(val);
         for i in 0..v.len(){
-            self.store8(addr+i as u32,v[i]);
+            self.store8(addr+i,v[i]);
         }
     }
-    fn load32(&self, addr:u32)->Option<u32>{
+    fn load32(&self, addr:usize)->Option<u32>{
         if let Some(b0) = self.load8(addr+0){
             if let Some(b1) = self.load8(addr+1){
                 if let Some(b2) = self.load8(addr+2){
@@ -30,10 +30,10 @@ pub trait Interconnect: InterconnectClone{
         }
         None
     }
-    fn store32(&mut self, addr:u32,val:u32){
+    fn store32(&mut self, addr:usize,val:u32){
         let v = u32_to_u8(val);
         for i in 0..v.len(){
-            self.store8(addr+i as u32,v[i]);
+            self.store8(addr+i,v[i]);
         }
     }
     fn as_any(&self) -> &dyn Any;
